Update and rename flappyborb.py to flappy.py
I removed a bunch of unnecessary lines and added in a line to do unit testing without running the game.
